+++ client-v2/src/data/reducer/editor.js
