+++ client-v2/src/integration/fetchJSON.js
