+++ client-v2/src/ui/editor/adaptiveCardEditor/index.js
