+++ client-v2/src/ui/experimental/actionDropdown.js
@@ -1,6 +1,6 @@
+import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { connect } from 'react-redux';
 
 class ActionDropdown extends React.Component {
     constructor(props, context) {
