+++ client-v2/src/ui/layout/sideBySide.js
