+++ client-v2/src/ui/shell/emulatorEditor/index.js
@@ -1,5 +1,5 @@
-import React from 'react';
 import { css } from 'glamor';
+import React from 'react';
 
 export default class EmulatorEditor extends React.Component {
     render() {
